General: create new component for write role
